# erlang-server/generated/priv/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Jenkins API clients generated from Swagger / Open API specification",
    "version" : "0.1.0",
    "title" : "Swaggy Jenkins",
    "contact" : {
      "name" : "Cliffano Subagio",
      "url" : "http://github.com/cliffano/swaggy-jenkins",
      "email" : "blah@cliffano.com"
    }
  },
  "basePath" : "/",
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/api/json" : {
      "get" : {
        "tags" : [ "remoteAccess" ],
        "description" : "Retrieve Jenkins details",
        "operationId" : "getJenkins",
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Jenkins details",
            "schema" : {
              "$ref" : "#/definitions/hudsonmodelHudson"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          }
        }
      },
      "head" : {
        "tags" : [ "remoteAccess" ],
        "description" : "Retrieve Jenkins headers",
        "operationId" : "headJenkins",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved Jenkins headers",
            "headers" : {
              "x-jenkins" : {
                "type" : "string",
                "description" : "Jenkins version number"
              }
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          }
        }
      }
    },
    "/computer/api/json?depth=1" : {
      "get" : {
        "tags" : [ "remoteAccess" ],
        "description" : "Retrieve computer details",
        "operationId" : "getComputer",
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved computer details",
            "schema" : {
              "$ref" : "#/definitions/hudsonmodelComputerSet"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          }
        }
      }
    },
...
    "/createView" : {
      "post" : {
        "tags" : [ "remoteAccess" ],
        "description" : "Create a new view using view configuration",
        "operationId" : "postCreateView",
        "produces" : [ "text/html" ],
        "parameters" : [ {
          "name" : "name",
          "in" : "query",
          "description" : "Name of the new view",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "View configuration in config.xml format",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Jenkins-Crumb",
          "in" : "header",
          "description" : "CSRF protection token",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Content-Type",
          "in" : "header",
          "description" : "Content type header application/xml",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully created the view"
          },
          "400" : {
            "description" : "An error has occurred - error message is embedded inside the HTML response",
            "schema" : {
              "type" : "string"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          }
        }
      }
    },
    "/crumbIssuer/api/json" : {
      "get" : {
        "tags" : [ "remoteAccess" ],
        "description" : "Retrieve CSRF protection token",
        "operationId" : "getCrumb",
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved CSRF protection token",
            "schema" : {
              "$ref" : "#/definitions/hudsonsecuritycsrfDefaultCrumbIssuer"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          }
        }
      }
    },
    "/job/{name}/api/json" : {
      "get" : {
        "tags" : [ "remoteAccess" ],
        "description" : "Retrieve job details",
        "operationId" : "getJob",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "description" : "Name of the job",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved job details",
            "schema" : {
              "$ref" : "#/definitions/hudsonmodelFreeStyleProject"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          },
          "404" : {
            "description" : "Job cannot be found on Jenkins instance"
          }
        }
      }
    },
...
    "/job/{name}/lastBuild/api/json" : {
      "get" : {
        "tags" : [ "remoteAccess" ],
        "description" : "Retrieve job's last build details",
        "operationId" : "getJobLastBuild",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "description" : "Name of the job",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved job's last build details",
            "schema" : {
              "$ref" : "#/definitions/hudsonmodelFreeStyleBuild"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          },
          "404" : {
            "description" : "Job cannot be found on Jenkins instance"
          }
        }
      }
    },
...
    "/queue/api/json" : {
      "get" : {
        "tags" : [ "remoteAccess" ],
        "description" : "Retrieve queue details",
        "operationId" : "getQueue",
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved queue details",
            "schema" : {
              "$ref" : "#/definitions/hudsonmodelQueue"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          }
        }
      }
    },
    "/queue/item/{number}/api/json" : {
      "get" : {
        "tags" : [ "remoteAccess" ],
        "description" : "Retrieve queued item details",
        "operationId" : "getQueueItem",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "number",
          "in" : "path",
          "description" : "Queue number",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved queued item details",
            "schema" : {
              "$ref" : "#/definitions/hudsonmodelQueue"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          }
        }
      }
    },
    "/view/{name}/api/json" : {
      "get" : {
        "tags" : [ "remoteAccess" ],
        "description" : "Retrieve view details",
        "operationId" : "getView",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "description" : "Name of the view",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved view details",
            "schema" : {
              "$ref" : "#/definitions/hudsonmodelListView"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          },
          "404" : {
            "description" : "View cannot be found on Jenkins instance"
          }
        }
      }
    },
    "/view/{name}/config.xml" : {
      "get" : {
        "tags" : [ "remoteAccess" ],
        "description" : "Retrieve view configuration",
        "operationId" : "getViewConfig",
        "produces" : [ "text/xml" ],
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "description" : "Name of the view",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved view configuration in config.xml format",
            "schema" : {
              "type" : "string"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          },
          "404" : {
            "description" : "View cannot be found on Jenkins instance"
          }
        }
      },
      "post" : {
        "tags" : [ "remoteAccess" ],
        "description" : "Update view configuration",
        "operationId" : "postViewConfig",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "description" : "Name of the view",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "View configuration in config.xml format",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Jenkins-Crumb",
          "in" : "header",
          "description" : "CSRF protection token",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully updated view configuration"
          },
          "400" : {
            "description" : "An error has occurred - error message is embedded inside the HTML response",
            "schema" : {
              "type" : "string"
            }
          },
          "401" : {
            "description" : "Authentication failed - incorrect username and/or password"
          },
          "403" : {
            "description" : "Jenkins requires authentication - please set username and password"
          },
          "404" : {
            "description" : "View cannot be found on Jenkins instance"
          }
        }
      }
    },
...
  },
  "securityDefinitions" : {
    "jenkins_auth" : {
      "type" : "basic"
    }
  },
  "definitions" : {
    "hudsonmodelFreeStyleBuild" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        },
        "number" : {
          "type" : "integer"
        },
        "url" : {
          "type" : "string"
        },
        "actions" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/hudsonmodelCauseAction"
          }
        },
        "building" : {
          "type" : "boolean"
        },
        "description" : {
          "type" : "string"
        },
        "displayName" : {
          "type" : "string"
        },
        "duration" : {
          "type" : "integer"
        },
        "estimatedDuration" : {
          "type" : "integer"
        },
        "executor" : {
          "type" : "string"
        },
        "fullDisplayName" : {
          "type" : "string"
        },
        "id" : {
          "type" : "string"
        },
        "keepLog" : {
          "type" : "boolean"
        },
        "queueId" : {
          "type" : "integer"
        },
        "result" : {
          "type" : "string"
        },
        "timestamp" : {
          "type" : "integer"
        },
        "builtOn" : {
          "type" : "string"
        },
        "changeSet" : {
          "$ref" : "#/definitions/hudsonscmEmptyChangeLogSet"
        }
      }
    },
    "hudsonmodelHudsonMasterComputerexecutors" : {
      "type" : "object",
      "properties" : {
        "currentExecutable" : {
          "$ref" : "#/definitions/hudsonmodelFreeStyleBuild"
        },
        "idle" : {
          "type" : "boolean"
        },
        "likelyStuck" : {
          "type" : "boolean"
        },
        "number" : {
          "type" : "integer"
        },
        "progress" : {
          "type" : "integer"
        },
        "_class" : {
          "type" : "string"
        }
      }
    },
    "hudsonmodelLabel1" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        }
      }
    },
    "hudsonnode_monitorsSwapSpaceMonitorMemoryUsage2" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        },
        "availablePhysicalMemory" : {
          "type" : "integer"
        },
        "availableSwapSpace" : {
          "type" : "integer"
        },
        "totalPhysicalMemory" : {
          "type" : "integer"
        },
        "totalSwapSpace" : {
          "type" : "integer"
        }
      }
    },
    "hudsonnode_monitorsDiskSpaceMonitorDescriptorDiskSpace" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        },
        "timestamp" : {
          "type" : "integer"
        },
        "path" : {
          "type" : "string"
        },
        "size" : {
          "type" : "integer"
        }
      }
    },
...
    "hudsonmodelHudsonMasterComputer_monitorData" : {
      "type" : "object",
      "properties" : {
        "hudson.node_monitors.SwapSpaceMonitor" : {
          "$ref" : "#/definitions/hudsonnode_monitorsSwapSpaceMonitorMemoryUsage2"
        },
        "hudson.node_monitors.TemporarySpaceMonitor" : {
          "$ref" : "#/definitions/hudsonnode_monitorsDiskSpaceMonitorDescriptorDiskSpace"
        },
        "hudson.node_monitors.DiskSpaceMonitor" : {
          "$ref" : "#/definitions/hudsonnode_monitorsDiskSpaceMonitorDescriptorDiskSpace"
        },
        "hudson.node_monitors.ArchitectureMonitor" : {
          "type" : "string"
        },
        "hudson.node_monitors.ResponseTimeMonitor" : {
          "$ref" : "#/definitions/hudsonnode_monitorsResponseTimeMonitorData"
        },
        "hudson.node_monitors.ClockMonitor" : {
          "$ref" : "#/definitions/hudsonutilClockDifference"
        },
        "_class" : {
          "type" : "string"
        }
      }
    },
    "hudsonmodelHudsonMasterComputer" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        },
        "displayName" : {
          "type" : "string"
        },
        "executors" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/hudsonmodelHudsonMasterComputerexecutors"
          }
        },
        "icon" : {
          "type" : "string"
        },
        "iconClassName" : {
          "type" : "string"
        },
        "idle" : {
          "type" : "boolean"
        },
        "jnlpAgent" : {
          "type" : "boolean"
        },
        "launchSupported" : {
          "type" : "boolean"
        },
        "loadStatistics" : {
          "$ref" : "#/definitions/hudsonmodelLabel1"
        },
        "manualLaunchAllowed" : {
          "type" : "boolean"
        },
        "monitorData" : {
          "$ref" : "#/definitions/hudsonmodelHudsonMasterComputer_monitorData"
        },
        "numExecutors" : {
          "type" : "integer"
        },
        "offline" : {
          "type" : "boolean"
        },
        "offlineCause" : {
          "type" : "string"
        },
        "offlineCauseReason" : {
          "type" : "string"
        },
        "temporarilyOffline" : {
          "type" : "boolean"
        }
      }
    },
    "hudsonmodelComputerSet" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        },
        "busyExecutors" : {
          "type" : "integer"
        },
        "computer" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/hudsonmodelHudsonMasterComputer"
          }
        },
        "displayName" : {
          "type" : "string"
        },
        "totalExecutors" : {
          "type" : "integer"
        }
      }
    },
    "hudsonsecuritycsrfDefaultCrumbIssuer" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        },
        "crumb" : {
          "type" : "string"
        },
        "crumbRequestField" : {
          "type" : "string"
        }
      }
    },
    "hudsonmodelHudsonassignedLabels" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        }
      }
    },
    "hudsonmodelFreeStyleProject" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "url" : {
          "type" : "string"
        },
        "color" : {
          "type" : "string"
        },
        "actions" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/hudsonmodelFreeStyleProjectactions"
          }
        },
        "description" : {
          "type" : "string"
        },
        "displayName" : {
          "type" : "string"
        },
        "displayNameOrNull" : {
          "type" : "string"
        },
        "fullDisplayName" : {
          "type" : "string"
        },
        "fullName" : {
          "type" : "string"
        },
        "buildable" : {
          "type" : "boolean"
        },
        "builds" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/hudsonmodelFreeStyleBuild"
          }
        },
        "firstBuild" : {
          "$ref" : "#/definitions/hudsonmodelFreeStyleBuild"
        },
        "healthReport" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/hudsonmodelFreeStyleProjecthealthReport"
          }
        },
        "inQueue" : {
          "type" : "boolean"
        },
        "keepDependencies" : {
          "type" : "boolean"
        },
        "lastBuild" : {
          "$ref" : "#/definitions/hudsonmodelFreeStyleBuild"
        },
        "lastCompletedBuild" : {
          "$ref" : "#/definitions/hudsonmodelFreeStyleBuild"
        },
        "lastFailedBuild" : {
          "type" : "string"
        },
        "lastStableBuild" : {
          "$ref" : "#/definitions/hudsonmodelFreeStyleBuild"
        },
        "lastSuccessfulBuild" : {
          "$ref" : "#/definitions/hudsonmodelFreeStyleBuild"
        },
        "lastUnstableBuild" : {
          "type" : "string"
        },
        "lastUnsuccessfulBuild" : {
          "type" : "string"
        },
        "nextBuildNumber" : {
          "type" : "integer"
        },
        "queueItem" : {
          "type" : "string"
        },
        "concurrentBuild" : {
          "type" : "boolean"
        },
        "scm" : {
          "$ref" : "#/definitions/hudsonscmNullSCM"
        }
      }
    },
    "hudsonmodelAllView" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "url" : {
          "type" : "string"
        }
      }
    },
    "jenkinsmodelUnlabeledLoadStatistics" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        }
      }
    },
    "hudsonmodelHudson" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        },
        "assignedLabels" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/hudsonmodelHudsonassignedLabels"
          }
        },
        "mode" : {
          "type" : "string"
        },
        "nodeDescription" : {
          "type" : "string"
        },
        "nodeName" : {
          "type" : "string"
        },
        "numExecutors" : {
          "type" : "integer"
        },
        "description" : {
          "type" : "string"
        },
        "jobs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/hudsonmodelFreeStyleProject"
          }
        },
        "primaryView" : {
          "$ref" : "#/definitions/hudsonmodelAllView"
        },
        "quietingDown" : {
          "type" : "boolean"
        },
        "slaveAgentPort" : {
          "type" : "integer"
        },
        "unlabeledLoad" : {
          "$ref" : "#/definitions/jenkinsmodelUnlabeledLoadStatistics"
        },
        "useCrumbs" : {
          "type" : "boolean"
        },
        "useSecurity" : {
          "type" : "boolean"
        },
        "views" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/hudsonmodelAllView"
          }
        }
      }
    },
    "hudsonmodelFreeStyleProjectactions" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        }
      }
    },
    "hudsonmodelFreeStyleProjecthealthReport" : {
      "type" : "object",
      "properties" : {
        "description" : {
          "type" : "string"
        },
        "iconClassName" : {
          "type" : "string"
        },
        "iconUrl" : {
          "type" : "string"
        },
        "score" : {
          "type" : "integer"
        },
        "_class" : {
          "type" : "string"
        }
      }
    },
    "hudsonscmNullSCM" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        }
      }
    },
    "hudsonmodelCauseUserIdCause" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        },
        "shortDescription" : {
          "type" : "string"
        },
        "userId" : {
          "type" : "string"
        },
        "userName" : {
          "type" : "string"
        }
      }
    },
    "hudsonmodelCauseAction" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        },
        "causes" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/hudsonmodelCauseUserIdCause"
          }
        }
      }
    },
    "hudsonscmEmptyChangeLogSet" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        },
        "kind" : {
          "type" : "string"
        }
      }
    },
    "hudsonmodelQueueBlockedItem" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        },
        "actions" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/hudsonmodelCauseAction"
          }
        },
        "blocked" : {
          "type" : "boolean"
        },
        "buildable" : {
          "type" : "boolean"
        },
        "id" : {
          "type" : "integer"
        },
        "inQueueSince" : {
          "type" : "integer"
        },
        "params" : {
          "type" : "string"
        },
        "stuck" : {
          "type" : "boolean"
        },
        "task" : {
          "$ref" : "#/definitions/hudsonmodelFreeStyleProject"
        },
        "url" : {
          "type" : "string"
        },
        "why" : {
          "type" : "string"
        },
        "buildableStartMilliseconds" : {
          "type" : "integer"
        }
      }
    },
    "hudsonmodelQueue" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        },
        "items" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/hudsonmodelQueueBlockedItem"
          }
        }
      }
    },
...
    "hudsonmodelListView" : {
      "type" : "object",
      "properties" : {
        "_class" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "jobs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/hudsonmodelFreeStyleProject"
          }
        },
        "name" : {
          "type" : "string"
        },
        "url" : {
          "type" : "string"
        }
      }
    },
...
  },
...
}